add get_list_sorted(file)
