+++ kursovaya-obr/operations.json
